更进 WIDA 8.8 的两处修补：getfun 返回 T、mypow 改 i64
- 两点式转一般式 getfun 返回 tuple<int,int,int> → tuple<T,T,T>：T=long long 时 A/B/C 不再被截断（WIDA 2024.10.21）
- mypow 参数与累乘器 int → i64：默认模数 998244353 下 n*n 也曾溢出 int（WIDA 2024.10.15）；基础算法章溢出提示同步改为 p ≤ 3e9 安全

8.8 其余项核对后未跟：点到圆切线公式我们已先修过（base±acos(r/d)，与 8.8 修法等价）；割边缩点 10.20 重构 EmptyDust 已带上，现与 8.8 一致；蒙哥马利/卡常章、章节搬家与重构不收。mypow 对 __int128 参考值、getfun 用 2e9/3e9 坐标实测通过

+++ code/基础算法/快速幂与常用函数.cpp
@@ -13,8 +13,8 @@ const double eps = 1e-8;
 const double PI = acos(-1.0);
 
 // @book-begin
-int mypow(int n, int k, int p = MOD) {  // 快速幂，复杂度 O(log k)
-    int r = 1;
+int mypow(i64 n, i64 k, int p = MOD) {  // 快速幂，复杂度 O(log k)；参数与累乘器 i64，n*n 不溢出 int
+    i64 r = 1;
     for (; k; k >>= 1, n = n * n % p) {
         if (k & 1) r = r * n % p;
     }
